+++ src/GitCandy.Data.SqlServer/GitCandySqlServerDataBuilderExtensions.cs
@@ -2,32 +2,31 @@ using GitCandy.Data;
 using GitCandy.Data.Configuration;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using SonnetDB.EntityFrameworkCore.Extensions;
 
-namespace GitCandy.Data.SonnetDB;
+namespace GitCandy.Data.SqlServer;
 
 /// <summary>
-/// SonnetDB provider 注册扩展。
+/// SQL Server provider 注册扩展。
 /// </summary>
-public static class GitCandySonnetDbDataBuilderExtensions
+public static class GitCandySqlServerDataBuilderExtensions
 {
-    private const string MigrationsAssembly = "GitCandy.Data.SonnetDB";
+    private const string MigrationsAssembly = "GitCandy.Data.SqlServer";
 
     /// <summary>
-    /// 注册 SonnetDB provider。
+    /// 注册 SQL Server provider。
     /// </summary>
     /// <param name="builder">数据库构建器。</param>
     /// <returns>数据库构建器。</returns>
-    public static GitCandyDatabaseBuilder AddSonnetDB(this GitCandyDatabaseBuilder builder)
+    public static GitCandyDatabaseBuilder AddSqlServer(this GitCandyDatabaseBuilder builder)
     {
         ArgumentNullException.ThrowIfNull(builder);
 
-        return builder.AddProvider(GitCandyDatabaseProvider.SonnetDB, static (services, options) =>
+        return builder.AddProvider(GitCandyDatabaseProvider.SqlServer, static (services, options) =>
         {
             services.AddPooledDbContextFactory<GitCandyDbContext>(dbContextOptions =>
             {
-                dbContextOptions.UseSonnetDB(options.ConnectionString, sonnetOptions =>
-                    sonnetOptions.UseQuerySplittingBehavior(QuerySplittingBehavior.SplitQuery)
+                dbContextOptions.UseSqlServer(options.ConnectionString, sqlServerOptions =>
+                    sqlServerOptions.UseQuerySplittingBehavior(QuerySplittingBehavior.SplitQuery)
                         .MigrationsAssembly(MigrationsAssembly));
 
                 dbContextOptions.ConfigureGitCandyCommonOptions(options);